changes position from decimal to float

+++ database/migrations/2020_02_03_131611_create_tasks_table.php
@@ -16,7 +16,7 @@ class CreateTasksTable extends Migration
         Schema::create('tasks', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('title');
-            $table->decimal('position', 10, 8);
+            $table->float('position', 10, 8);
             $table->timestamps();
         });
     }
